continued work on vertex array
Also modified the buffer object and vertex array to use shared_ptr and
made them copyable. The shared_ptr will allow more efficient use of
already uploaded data on the GPU buffers that multiple objects of the
same type can share, and only when the final owner of that data is
deleted will it finally be removed from the GPU. However the data can
still be manually removed with a call to the erase function in either
object.